requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
requests==2.31.0
cloudflare-worker==0.1.1
werkzeug==3.0.1
